+++ platform/common/default.nix
@@ -1,8 +1,7 @@
 # common
 # shared stuff between linux and darwin
-{...} :
-{
+{ ... }: {
   imports = [
     # ./ssh.nix
-    ];
+  ];
 }