+++ platform/common/ssh.nix
@@ -1,13 +1,12 @@
 # shared ssh config between linux/MacOS, etc...
 # TODO:
 # USE ONEKEY FOR SSH
-{flake, outputs, pkgs, config, ...} :
+{ flake, outputs, pkgs, config, ... }:
 let
-# hosts = ["trantor" "terminus" "smyrno" "anacreon"];
-hosts = lib.attrNames outputs.nixosConfigurations;
-in
-{
-services.openssh = {
+  # hosts = ["trantor" "terminus" "smyrno" "anacreon"];
+  hosts = lib.attrNames outputs.nixosConfigurations;
+in {
+  services.openssh = {
     enable = true;
     settings = {
       # Harden
@@ -19,37 +18,32 @@ services.openssh = {
       GatewayPorts = "clientspecified";
     };
 
-    hostKeys = [
-      {
-        bits = 4096;
-        openSSHFormat = true;
-        path = "/etc/ssh/ssh_host_ed25519_key";
-        type = "ed25519";
-      }
-    ];
+    hostKeys = [{
+      bits = 4096;
+      openSSHFormat = true;
+      path = "/etc/ssh/ssh_host_ed25519_key";
+      type = "ed25519";
+    }];
   };
 
   programs.ssh = {
     # Each hosts public key
     knownHosts = lib.genAttrs hosts (hostname: {
-      publicKeyFile = ${flake}/systems/${hostname}/ssh_host_ed25519_key.pub;
-      extraHostNames = [ "${hostname}.${config.}"
-        ]
-        ++
+      publicKeyFile = "${flake}/systems/${hostname}/ssh_host_ed25519_key.pub";
+      extraHostNames = [ "${hostname}.${config.networking.domain}" ] ++
         # Alias for localhost if it's the same host
         (lib.optional (hostname == config.networking.hostName) "localhost");
     });
   };
 
   # Passwordless sudo when SSH'ing with keys
-  security.pam.services.sudo = {config, ...}: {
+  security.pam.services.sudo = { config, ... }: {
     rules.auth.rssh = {
       order = config.rules.auth.ssh_agent_auth.order - 1;
       control = "sufficient";
       modulePath = "${pkgs.pam_rssh}/lib/libpam_rssh.so";
       settings.authorized_keys_command =
-        pkgs.writeShellScript "get-authorized-keys"
-        ''
+        pkgs.writeShellScript "get-authorized-keys" ''
           cat "/etc/ssh/authorized_keys.d/$1"
         '';
     };