+++ platform/linux/core/disks/default.nix
@@ -1,9 +1,5 @@
 # core/disks/default.nix
 # 	Nixos disks and partition setup
 { ... }: {
-  imports = [
-    #./btrfs.nix
-    ./tools.nix
-    ./zfs.nix
-  ];
+  imports = [ ./btrfs.nix ./tools.nix ./zfs.nix ./samba.nix ];
 }