+++ platform/linux/core/disks/samba.nix
@@ -1,0 +1,5 @@
+# disks/samba.nix
+# settings to detect and mount samba shares
+{ ... }: {
+  services.samba-wsdd.workgroup = "WORKGROUP";
+}