+++ systems/NUC/default.nix
@@ -1,10 +1,6 @@
 # NUC
 #   this is a 12th gen Intel NUC
 #   it's my central Home Cloud
-{  ... }:
-{
-  imports = [
-    ./configuration.nix
-    ./server.nix
-  ];
+{ ... }: {
+  imports = [ ./configuration.nix ./server.nix ];
 }