+++ systems/SCP/default.nix
@@ -1,6 +1,5 @@
 # SCP
 # Samsung Chromebook Pro (Caroline)
-{...} :
-{
-  imports = [./configuration.nix];
+{ ... }: {
+  imports = [ ./configuration.nix ];
 }