+++ systems/TAF/default.nix
@@ -2,8 +2,5 @@
 #   previously "AF"
 #   this is my main desktop PC
 { pkgs, ... }: {
-  imports = [
-    ./configuration.nix
-    ./hardware-configuration.nix
-    ];
+  imports = [ ./configuration.nix ./hardware-configuration.nix ];
 }