+++ users/perard/home/applications/default.nix
@@ -2,8 +2,8 @@
 # TODO : make this a shareable module (between users)
 { ... }: {
   imports = [
+    ./discord
     ./deezer.nix
-    ./discord.nix
     ./firefox.nix
     ./games.nix
     ./multimedia.nix